move TestInit into init_test.go file
as requested by pd93.
