+++ task_test.go
@@ -1011,27 +1011,6 @@ func TestCmdsVariables(t *testing.T) {
 	assert.Contains(t, buff.String(), tf)
 }
 
-func TestInit(t *testing.T) {
-	t.Parallel()
-
-	const dir = "testdata/init"
-	file := filepathext.SmartJoin(dir, "Taskfile.yml")
-
-	_ = os.Remove(file)
-	if _, err := os.Stat(file); err == nil {
-		t.Errorf("Taskfile.yml should not exist")
-	}
-
-	if err := task.InitTaskfile(io.Discard, dir); err != nil {
-		t.Error(err)
-	}
-
-	if _, err := os.Stat(file); err != nil {
-		t.Errorf("Taskfile.yml should exist")
-	}
-	_ = os.Remove(file)
-}
-
 func TestCyclicDep(t *testing.T) {
 	t.Parallel()
